requirements.txt
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
openai==1.0.0
tqdm==4.65.0
pyyaml==6.0
rich==13.6.0
pytest==7.4.0